+++ src/app/components/NavLink.tsx
@@ -20,7 +20,6 @@ interface NavLinksProps {
 
 // Map of links to display in the side navigation.
 const links = [
-  { name: "Home", href: "/home", icon: HomeIcon },
   {
     name: "Cadastrar raça",
     href: "/administracao/cadastroRaca",
@@ -39,32 +38,10 @@ export default function NavLinks({ minimized = false }: NavLinksProps) {
   const pathname = usePathname();
   return (
     <>
-    {links.slice(0, 1).map((link) => {
-        const LinkIcon = link.icon;
-        return (
-          <Link
-            key={link.name}
-            href={link.href}
-            className={clsx(
-              "flex h-[48px] grow items-center justify-center gap-2 rounded-md bg-purple-0 p-3 text-sm font-medium text-gray-400 hover:bg-purple-100 hover:text-purple-800 md:flex-none md:justify-start md:p-2 md:px-3",
-              {
-                "bg-purple-200 text-purple-800": pathname === link.href,
-              }
-            )}
-          >
-            <LinkIcon className="w-6" />
-            {!minimized && (
-          <p className="hidden md:block">
-            {link.name}
-          </p>
-        )}
-          </Link>
-        );
-      })}
       {!minimized && (
       <h3 className="px-3 py-2 text-gray-400">ADMINISTRAÇÃO</h3>
       )}
-      {links.slice(1, 5).map((link) => {
+      {links.slice(0, 5).map((link) => {
         const LinkIcon = link.icon;
         return (
           <Link
